pyproject.toml
[project]
name = "fantasy_langgen"
version = "0.3.0"
description = ""
authors = [
    {name = "Jody Huyck-Jones",email = "68274088+jdhuyck@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
flake8 = "^7.3.0"
isort = "^6.0.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.16"
check-jsonschema = "^0.33.2"

